백지영 - 게시글 수정시 첨부파일 삭제 문제 해결 29일 14:39

+++ src/com/kh/elefont/community/controller/CommunityUpdateServlet.java
@@ -90,15 +90,13 @@ public class CommunityUpdateServlet extends HttpServlet {
 			
 			System.out.println(community);
 			// 첨부파일
-			//File f = multipartRequest.getFile("upFile");
-//			if(f != null) {
-				Attachment attach = new Attachment();
-				attach.setMemberNo(attach.getMemberNo());
-				attach.setCommNo(commNo); 
-				attach.setOriginalFilename(multipartRequest.getOriginalFileName("upFile"));
-				attach.setRenamedFilename(multipartRequest.getFilesystemName("upFile"));
-				community.setAttach(attach);
-				
+			File f = multipartRequest.getFile("upFile");
+			Attachment attach = new Attachment();
+			attach.setMemberNo(attach.getMemberNo());
+			attach.setCommNo(commNo); 
+			attach.setOriginalFilename(multipartRequest.getOriginalFileName("upFile"));
+			attach.setRenamedFilename(multipartRequest.getFilesystemName("upFile"));
+			community.setAttach(attach);
 			
 			
 			System.out.println("community@servlet = " + community);
@@ -106,21 +104,22 @@ public class CommunityUpdateServlet extends HttpServlet {
 			// 2. 업무로직 
 			int result = 0;
 //			// 기존파일 삭제 (서버컴퓨터 파일 삭제 + db 레코드삭제)
-			Attachment oldattach = attachmentService.selectOneAttachment(commNo);
-//			// 서버컴퓨터 파일 삭제
-			File _delFile = new File(saveDirectory, oldattach.getRenamedFilename());
-			_delFile.delete();
-//			// db 레코드삭제
-//			result = attachmentService.deleteAttachmentByCommNo(commNo);
-//			System.out.println(result > 0 ? "첨부파일 삭제 성공!" : "첨부파일 삭제 실패!");
-//			
+			String delFile = multipartRequest.getParameter("delFile");
+			if(delFile != null) {
+				Attachment oldattach = attachmentService.selectOneAttachment(commNo);
+				// 서버컴퓨터 파일 삭제
+				File _delFile = new File(saveDirectory, oldattach.getRenamedFilename());
+				_delFile.delete();
+	//			// db 레코드삭제
+	//			result = attachmentService.deleteAttachmentByCommNo(commNo);
+	//			System.out.println(result > 0 ? "첨부파일 삭제 성공!" : "첨부파일 삭제 실패!");
+			}
 			
 			// 게시물 수정 + 첨부파일 등록
 			result = attachmentService.updateAttachment(attach);
 			result = communityService.updateCommunity(community);
 			String msg = result > 0 ? "게시물 수정 성공!" : "게시물 수정 실패!";
 			
-			System.out.println("제발삭제되어라"+oldattach.getRenamedFilename());
 	
 			// 3. redirect
 			request.getSession().setAttribute("msg", msg);
